+++ package.json
@@ -14,13 +14,16 @@
   "bin": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --unhandled-rejections=strict src",
+    "start": "node --unhandled-rejections=strict --require ts-node/register src",
     "postpublish": "git push --follow-tags"
   },
   "devDependencies": {
     "@types/node": "13.9.0",
     "@types/node-fetch": "2.5.5",
-    "@types/yauzl": "2.9.1"
+    "@types/validate-npm-package-name": "3.0.0",
+    "@types/yauzl": "2.9.1",
+    "ts-node": "8.6.2",
+    "typescript": "3.8.3"
   },
   "dependencies": {
     "commander": "4.1.1",
