--- src/index.js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require("./create-package-javascript")();
