+++ src/index.ts
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require("./create-package-javascript").default();
