+++ src/program.ts
@@ -1,8 +1,16 @@
-const commander = require("commander");
+import commander from "commander";
 
-module.exports = commander
+/** @private */
+interface Program extends commander.Command {
+	in?: string;
+}
+
+/** @public */
+const program: Program = commander
 	.name("create-package-javascript")
 	.usage("<name> [--in <path>]")
 	.option("<name>", "valid package name\nsee https://github.com/npm/validate-npm-package-name#naming-rules")
 	.option("-p, --in <path>", "path for new package code\nrelative to process's current working directory")
 	.parse(process.argv);
+
+export default program;
